HPCC-34264: Cleanup after event visitation refactoring
