+++ common/eventconsumption/CMakeLists.txt
@@ -27,11 +27,10 @@ set (
     ${CMAKE_CURRENT_SOURCE_DIR}/eventdumpstream.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/eventdumpptree.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/eventfilter.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/eventindex.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/eventindexhotspot.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/eventindexhotspotbuckets.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/eventindexsummarize.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/eventoperation.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/eventvisitor.cpp
 )
 
 ADD_DEFINITIONS( -D_USRDLL -DEVENTCONSUMPTION_EXPORTS )
