+++ common/eventconsumption/eventdump.hpp
@@ -63,13 +63,13 @@ public: // IEventVisitor
             switch (attr.queryTypeClass())
             {
             case EATCtext:
-                visitAttribute(attr.queryId(), attr.queryTextValue());
+                doVisitAttribute(attr.queryId(), attr.queryTextValue());
                 break;
             case EATCnumeric:
-                visitAttribute(attr.queryId(), attr.queryNumericValue());
+                doVisitAttribute(attr.queryId(), attr.queryNumericValue());
                 break;
             case EATCboolean:
-                visitAttribute(attr.queryId(), attr.queryBooleanValue());
+                doVisitAttribute(attr.queryId(), attr.queryBooleanValue());
                 break;
             default:
                 throw makeStringExceptionV(-1, "unsupported attribute type class %u", attr.queryTypeClass());
@@ -81,21 +81,6 @@ public: // IEventVisitor
 
 protected:
     virtual void visitEvent(EventType id) {};
-
-    inline void visitAttribute(EventAttr id, const char* value)
-    {
-        doVisitAttribute(id, value);
-    }
-
-    inline void visitAttribute(EventAttr id, bool value)
-    {
-        doVisitAttribute(id, value);
-    }
-
-    inline void visitAttribute(EventAttr id, __uint64 value)
-    {
-        doVisitAttribute(id, value);
-    }
 
     virtual void departEvent() {};
