+++ common/eventconsumption/eventindexhotspot.cpp
@@ -16,186 +16,9 @@
 ############################################################################## */
 
 #include "eventindexhotspot.h"
-#include "eventindex.hpp"
+#include "eventindexhotspot.hpp"
 #include <array>
 #include <map>
-
-// Report all visited buckets, in order, with counts and bucket kinds. At most one bucket may be
-// ambiguous with respect to bucket kind, as only one bucket may include the final leaf and initial
-// branch pages.
-class CAllBucketVisitor : public CStreamingBucketVisitor
-{
-public:
-    virtual void visitBucket(bucket_type bucket, BucketKind bucketKind, stat_type stat) override
-    {
-        stats[bucketKind].recordEvent(stat);
-        StringBuffer lines;
-        if (firstLeaf && BucketLeaf == bucketKind)
-        {
-            lines.append("  leaf:\n");
-            firstLeaf = false;
-        }
-        else if (firstBranch && BucketBranch == bucketKind)
-        {
-            lines.append("  branch:\n");
-            firstBranch = false;
-        }
-        lines.append("  - first-node: ").append(bucket2page(bucket, currentGranularity)).append('\n');
-        lines.append("    events: ").append(stat).append('\n');
-        out->put(lines.length(), lines.str());
-    }
-
-    virtual void depart() override
-    {
-        StringBuffer lines;
-        stats[BucketLeaf].toString(lines, 2, "leafStats");
-        stats[BucketAmbiguous].toString(lines, 2, "ambiguousStats");
-        stats[BucketBranch].toString(lines, 2, "branchStats");
-        out->put(lines.length(), lines.str());
-        firstLeaf = firstBranch = true;
-    }
-
-    virtual bool wantAmbiguous() const override
-    {
-        return false;
-    }
-
-protected:
-    virtual void appendAnalysis(StringBuffer& lines) const override
-    {
-        lines.append("all buckets");
-    }
-
-public:
-    using CStreamingBucketVisitor::CStreamingBucketVisitor;
-
-private:
-    bool firstLeaf{true};
-    bool firstBranch{true};
-};
-
-// Report up to the N buckets of each bucket kind with the highest activity counts. Buckets with
-// the same activity count as the Nth bucket are reported cumulatively, with the number of buckets
-// in place of individual bucket numbers.
-class CTopBucketVisitor : public CStreamingBucketVisitor
-{
-protected:
-    using Hit = std::pair<stat_type, bucket_type>;
-    using Hits = std::vector<Hit>;
-    struct Info
-    {
-        Hits hits;
-        stat_type firstDrop;
-        __uint64 droppedCount;
-
-        Info() { clear(); }
-        bool empty() const { return hits.empty(); }
-
-        void clear()
-        {
-            hits.clear();
-            firstDrop = 0;
-            droppedCount = 0;
-        }
-    };
-
-public:
-    virtual void visitBucket(bucket_type bucket, BucketKind bucketKind, stat_type stat) override
-    {
-        Info& info = (BucketLeaf == bucketKind ? leaves : branches);
-        stats[bucketKind].recordEvent(stat);
-        if (info.hits.size() == limit) // have top N; drop the lowest
-        {
-            stat_type last = info.hits.back().first;
-            if (stat < last) // ignore values lower than the lowest
-                return;
-            if (stat == last) // drop a value equal to the lowest
-            {
-                if (info.firstDrop != stat)
-                    info.firstDrop = stat;
-                info.droppedCount++;
-                return;
-            }
-            info.hits.pop_back(); // drop the last value lower than the new one
-
-            if (info.hits.back().first == last)
-            {
-                if (last > info.firstDrop) // dropped the first occurence of the new lowest value
-                {
-                    info.firstDrop = last;
-                    info.droppedCount = 1;
-                }
-                else // dropped another occurrence of the lowest value
-                    info.droppedCount++;
-            }
-            else // no occurrences of the lowest value have been dropped yet
-            {
-                info.firstDrop = 0;
-                info.droppedCount = 0;
-            }
-        }
-        // insert a new top N value
-        Hit v(stat, bucket);
-        Hits::const_iterator it = std::lower_bound(info.hits.begin(), info.hits.end(), v, [](const Hit& a, const Hit& b) {
-            return a.first > b.first;
-        });
-        info.hits.insert(it, v);
-    }
-
-    virtual void depart() override
-    {
-        summarize(leaves, stats[BucketLeaf], "leaves");
-        summarize(branches, stats[BucketBranch], "branches");
-        CStreamingBucketVisitor::depart();
-    }
-
-    virtual bool wantAmbiguous() const override
-    {
-        // Output clearly differentiates between leaf and branch buckets. Ambiguity is not needed.
-        return false;
-    }
-
-protected:
-    virtual void appendAnalysis(StringBuffer& lines) const override
-    {
-        lines.append("top ").append(int(limit)).append(" buckets");
-    }
-
-public:
-    CTopBucketVisitor(IBufferedSerialOutputStream& _out, byte _limit)
-        : CStreamingBucketVisitor(_out)
-        , limit(_limit)
-    {
-    }
-
-protected:
-    void summarize(Info& info, Stats& stats, const char* label)
-    {
-        if (info.empty())
-            return;
-        StringBuffer lines;
-        lines.append("  ").append(label).append(":\n");
-        stats.toString(lines, 4);
-        lines.append("    bucket:\n");
-        for (const Hit& hit : info.hits)
-        {
-            lines.append("    - first-node: ").append(bucket2page(hit.second, currentGranularity)).append('\n');
-            lines.append("      events: ").append(hit.first).append('\n');
-        }
-        if (info.droppedCount)
-        {
-            lines.append("    - dropped: ").append(info.droppedCount).append('\n');
-            lines.append("      events: ").append(info.firstDrop).append('\n');
-        }
-        out->put(lines.length(), lines.str());
-        info.clear();
-    }
-
-private:
-    Info leaves;
-    Info branches;
-    byte limit;
-};
 
 class CHotspotEventVisitor : public CInterfaceOf<IEventVisitor>
 {
@@ -321,7 +144,7 @@ class CHotspotEventVisitor : public CInterfaceOf<IEventVisitor>
         Activity activity[BucketAmbiguous]; // incoming activity is never ambiguous, this storage is not ambiguous
     };
 
-public: // IEventAttributeVisitor
+public: // IEventVisitor
     virtual bool visitFile(const char* filename, uint32_t version) override
     {
         return true;
@@ -378,9 +201,9 @@ bool CIndexHotspotOp::doOp()
 {
     Owned<IBucketVisitor> analyzer;
     if (limit)
-        analyzer.setown(new CTopBucketVisitor(*out, limit));
+        analyzer.setown(createTopBucketVisitor(*out, limit));
     else
-        analyzer.setown(new CAllBucketVisitor(*out));
+        analyzer.setown(createAllBucketVisitor(*out));
 
     CHotspotEventVisitor visitor(*analyzer, observedEvent, granularityBits);
     return traverseEvents(inputPath, visitor);
